requirements.txt
matplotlib==3.7.1
numpy==1.25.0
scikit-learn==1.2.2
scipy==1.10.1
einops==0.7.0
huggingface-hub==0.17.3
meshio==5.3.4
opencv-python==4.8.1.78
plyfile==1.0.1
torch==2.3.0
torchvision==0.18.0
xformers
transformers==4.34.1
trimesh==4.0.0
potpourri3d==1.0.0
robust_laplacian==0.2.7
wandb==0.16.6
tensorboard==2.13.0
Pillow==11.1.0
umap-learn
libigl
open3d
polyscope